requirements.txt
# Bot requirements
# discord.py is async and works well with aiohttp

discord.py>=2.4.0
python-dotenv>=1.0.1
aiohttp>=3.9.5
